Changes to the users package
I have created a User superclass and adapted the Student, TA & Professor classes to extend it
