--- bin/.scannerwork/report-task.txt
@@ -1,6 +1,0 @@
-projectKey=Maintenance-assignment
-serverUrl=http://localhost:9000
-serverVersion=9.4.0.54424
-dashboardUrl=http://localhost:9000/dashboard?id=Maintenance-assignment
-ceTaskId=AYD6op2xDJb8WjR4Rx__
-ceTaskUrl=http://localhost:9000/api/ce/task?id=AYD6op2xDJb8WjR4Rx__
