+++ src/Users/Professor.java
@@ -2,53 +2,18 @@ package users;
 
 import java.sql.Date;
 
-public class Professor {
-	String name;
+public class Professor extends User{
 	String department;
 	Date hirringDate;
-	String PhDTopic;
-	String email;
-	String phoneNumber;
+	String PhDTopic;	
 	
-	
-	public Professor(String name, String department, Date hirringDate, String phDTopic, String email,
-			String phoneNumber) {
+	public Professor( String department, Date hirringDate, String phDTopic) {
 		super();
-		this.name = name;
 		this.department = department;
 		this.hirringDate = hirringDate;
 		PhDTopic = phDTopic;
-		this.email = email;
-		this.phoneNumber = phoneNumber;
 	}
 	
-	
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
 	public String getDepartment() {
 		return department;
 	}
@@ -66,14 +31,5 @@ public class Professor {
 	}
 	public void setPhDTopic(String phDTopic) {
 		PhDTopic = phDTopic;
-	}
-	
-	
-	public void notifyProfessor(String message) {
-		// do some stuff
-		
-	}
-	
-	
-	
+	}	
 }
