+++ src/Users/Student.java
@@ -4,44 +4,11 @@ public class Student {
 	String name;
 	int Id;
 	String email;
-	String phoneNumber;
+	String phoneNumber;	
 	
-	
-	
-	public Student(String name, int id, String email, String phoneNumber) {
+	public Student(int id) {
 		super();
-		this.name = name;
 		Id = id;
-		this.email = email;
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
 	}
 
 	public int getId() {
@@ -50,11 +17,5 @@ public class Student {
 
 	public void setId(int id) {
 		Id = id;
-	}
-	
-	
-	public void notifyStudent(String message) {
-		// do some stuff
-	}
-	
+	}	
 }
